+++ Changelog.md
@@ -3,6 +3,7 @@
 
 **Fix**
 - `/trans/convertRawTxStr`接口返回未签名交易体的哈希值或签名后的交易体字符串
+- 修改发交易入参中的`funcParam`，从`List<Object>`改为`List<String>`
 
 **兼容性**
 - 支持FISCO-BCOS v3.0.0 及以上版本